Redesign P4 final: delete the legacy AppShell chrome
All eleven routes now own their chrome via <Workspace>, so the old
shell goes: AppShell.tsx/.css (top bar, drawer, portals, help popup),
ActionFloater.tsx/.css, useFloaterDrag.ts, and the long-dead
QuarterTurnBar. index.tsx is a bare hash router (gallery at #/ and as
the unknown-hash fallback); the MIGRATED interregnum set is gone.
AppDescriptor now lives in apps.ts. The unlisted #/fractals-cpu legacy
route sizes to its container and gains a minimal brand-mark Home
button. Readme.tsx (explainer modal) and ToggleMenu.tsx (Fractals2D)
survive.
