--- src/components/ActionFloater.css
@@ -1,109 +1,0 @@
-/* ActionFloater — draggable floating actions panel (default: bottom-right).
-   Collapsed: a play button with a drag grip on its left edge. */
-
-.af {
-  position: absolute;
-  bottom: 16px;
-  left: 16px;
-  z-index: 35;
-  background: rgba(12, 12, 16, 0.72);
-  backdrop-filter: blur(8px);
-  -webkit-backdrop-filter: blur(8px);
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: 10px;
-  box-shadow: 0 6px 20px rgba(0, 0, 0, 0.4);
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-}
-.af-inactive { display: none; }
-
-.af-collapsed { padding: 3px; }
-.af-open { padding: 10px; width: 240px; max-width: calc(100vw - 32px); }
-
-/* Drag grip — present in both states. */
-.af-grip {
-  color: #6b6b73;
-  font-size: 13px;
-  line-height: 1;
-  cursor: grab;
-  touch-action: none;
-  user-select: none;
-  padding: 4px 2px;
-}
-.af-grip:active { cursor: grabbing; }
-
-/* Collapsed: grip on the left edge + play button. */
-.af-collapsed-row {
-  display: flex;
-  align-items: center;
-  gap: 2px;
-}
-.af-grip-collapsed {
-  align-self: stretch;
-  display: flex;
-  align-items: center;
-  border-radius: 6px;
-}
-.af-grip-collapsed:hover { background: rgba(255, 255, 255, 0.06); }
-
-.af-play {
-  width: 34px;
-  height: 34px;
-  border-radius: 6px;
-  border: none;
-  background: transparent;
-  color: var(--cp-accent, #ffd400);
-  font-size: 14px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.af-play:hover { background: rgba(255, 255, 255, 0.08); }
-
-/* Open: draggable header. */
-.af-header {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  margin-bottom: 8px;
-}
-.af-title {
-  flex: 1;
-  font-size: 11px;
-  color: #9b9ba3;
-  font-weight: 700;
-  letter-spacing: 0.06em;
-  text-transform: uppercase;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-.af-iconbtn {
-  background: transparent;
-  border: none;
-  color: #9b9ba3;
-  cursor: pointer;
-  font-size: 16px;
-  width: 24px;
-  height: 24px;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-}
-.af-iconbtn:hover { background: rgba(255, 255, 255, 0.08); color: #f0f0f3; }
-
-/* The actions render here (portaled from ShellActions). Reuse the drawer's
-   section-body spacing so buttons stack the same way. */
-.af-body {
-  max-height: 60vh;
-  overflow-y: auto;
-  overscroll-behavior: contain;
-}
-.af-body .cp-section-body {
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
